Implement issue #5 folder architecture across backend and frontend
