+++ crates/sql-intelliscan-repository/src/data_access/mod.rs
@@ -1,0 +1,3 @@
+mod static_backend_metadata_repository;
+
+pub use static_backend_metadata_repository::StaticBackendMetadataRepository;
