+++ crates/sql-intelliscan-repository/src/data_access/static_backend_metadata_repository.rs
@@ -1,8 +1,6 @@
 use sql_intelliscan_common::backend_origin;
 
-pub trait BackendMetadataRepository {
-    fn origin(&self) -> &'static str;
-}
+use crate::contracts::BackendMetadataRepository;
 
 #[derive(Debug, Default, Clone, Copy)]
 pub struct StaticBackendMetadataRepository;
