+++ crates/sql-intelliscan-repository/src/errors/mod.rs
@@ -1,0 +1,3 @@
+mod repository_error;
+
+pub use repository_error::RepositoryError;
