+++ crates/sql-intelliscan-repository/src/errors/repository_error.rs
@@ -1,0 +1,4 @@
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum RepositoryError {
+    SourceUnavailable,
+}
