+++ crates/sql-intelliscan-repository/src/lib.rs
@@ -1,3 +1,13 @@
-mod repository;
+pub mod data_access;
+pub mod errors;
+pub mod models;
 
-pub use repository::{BackendMetadataRepository, StaticBackendMetadataRepository};
+pub use data_access::StaticBackendMetadataRepository;
+
+pub mod contracts {
+    pub trait BackendMetadataRepository {
+        fn origin(&self) -> &'static str;
+    }
+}
+
+pub use contracts::BackendMetadataRepository;
