+++ crates/sql-intelliscan-repository/src/models/backend_metadata.rs
@@ -1,0 +1,4 @@
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct BackendMetadata {
+    pub origin: &'static str,
+}
