+++ crates/sql-intelliscan-repository/src/models/mod.rs
@@ -1,0 +1,3 @@
+mod backend_metadata;
+
+pub use backend_metadata::BackendMetadata;
