+++ crates/sql-intelliscan-services/src/contracts/backend_metadata_repository_contract.rs
@@ -1,0 +1,1 @@
+pub use sql_intelliscan_repository::BackendMetadataRepository;
