+++ crates/sql-intelliscan-services/src/errors/mod.rs
@@ -1,0 +1,3 @@
+mod service_error;
+
+pub use service_error::ServiceError;
