+++ crates/sql-intelliscan-services/src/errors/service_error.rs
@@ -1,0 +1,4 @@
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum ServiceError {
+    InvalidName,
+}
