+++ crates/sql-intelliscan-services/src/models/greeting_model.rs
@@ -1,0 +1,4 @@
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct GreetingModel {
+    pub message: String,
+}
