+++ crates/sql-intelliscan-services/src/models/mod.rs
@@ -1,0 +1,3 @@
+mod greeting_model;
+
+pub use greeting_model::GreetingModel;
