+++ crates/sql-intelliscan-services/src/use_cases/mod.rs
@@ -1,0 +1,3 @@
+mod greet_use_case;
+
+pub use greet_use_case::{greet, GreetingService};
