+++ src-tauri/src/configuration/mod.rs
@@ -1,0 +1,3 @@
+mod app_builder;
+
+pub use app_builder::{build_app, run_builder};
