+++ src-tauri/src/dependency_wiring/mod.rs
@@ -1,0 +1,3 @@
+mod service_registry;
+
+pub use service_registry::greet_user;
