+++ src-tauri/src/dependency_wiring/service_registry.rs
@@ -1,0 +1,5 @@
+use sql_intelliscan_services::greet as greet_from_services;
+
+pub fn greet_user(name: &str) -> String {
+    greet_from_services(name)
+}
