+++ src/lib.rs
@@ -1,3 +1,5 @@
 pub mod app;
+pub mod components;
+pub mod services;
 
-pub use app::App;
+pub use components::app::App;
