+++ src/services/mod.rs
@@ -1,0 +1,2 @@
+pub mod greeting_service;
+pub mod tauri_client;
